fix: correct float params that were incorrectly cast to int

+++ autotrade/strategies/hot_money.py
@@ -42,10 +42,10 @@ class HotMoneyStrategy(Strategy):
     ):
         self.allowed_entry_dates = set(allowed_entry_dates or [])
         self.use_screener = use_screener and bool(self.allowed_entry_dates)
-        self.trailing_activate = int(trailing_activate)
+        self.trailing_activate = float(trailing_activate)
         self.trailing_drawdown = float(trailing_drawdown)
         self.time_stop_days = int(time_stop_days)
-        self.time_stop_min_gain = int(time_stop_min_gain)
+        self.time_stop_min_gain = float(time_stop_min_gain)
         self.stop_loss = float(stop_loss)
         self.trend_ma_fast = int(trend_ma_fast)
         self.trend_ma_mid = int(trend_ma_mid)
